server/fed-learn-server-aggregator/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==43.0.3
fastapi==0.119.0
flwr==1.15.2
grpcio==1.75.1
h11==0.16.0
httptools==0.7.1
idna==3.11
iterators==0.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mysql-connector-python==9.2.0
numpy==2.1.3
pathspec==0.12.1
protobuf==4.25.8
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==13.9.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tomli==2.3.0
tomli_w==1.2.0
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wheel==0.45.1
